net/mlx5e: Extract mod hdr entry lookup/create code into function
To remove dependency on rtnl lock, code that looks up mod header entry and
creates new one if entry is not found, must do this atomically to be safe
when concurrent updates are possible. Extract code that looks up or creates
new mod header entry into standalone function as a preparation for
described refactoring.

issue: 1521224
Change-Id: I0cbc15c05647b4f2902599b9ab3bd9852f1d30fb
Signed-off-by: Vlad Buslov <vladbu@mellanox.com>