net/mlx5e: Extract hairpin entry lookup/create code into function
To remove dependency on rtnl lock, code that looks up hairpin entry and
creates new one if entry is not found, must do this atomically to be safe
when concurrent updates are possible. Extract code that looks up or creates
new hairpin entry into standalone function as a preparation for described
refactoring.

issue: 1521224
Change-Id: Ia79fa26eb0e8359fdb2aa66a334732b0f41f3f24
Signed-off-by: Vlad Buslov <vladbu@mellanox.com>