net: sched: flower: track filter deletion with flag
In order to prevent double deletion of filter by concurrent tasks when rtnl
lock is not used for synchronization, add 'deleted' filter flag. Check
for this flag when modifying filters and return error if concurrent
deletion is detected.

Refactor __fl_delete() to accept pointer to 'last' boolean as argument,
and return error code as function return value instead. This is necessary
to signal concurrent filter delete to caller.

Change-Id: If79e1f00916cd96ced2a20393eb23138a45d7616
Signed-off-by: Vlad Buslov <vladbu@mellanox.com>