fixup! net: sched: refactor tp insert/delete for concurrent execution
When same prio protos are created concurrently, insertion of one of them
fails and proto is destroyed. In this case 'tp_created' flag is not set and
chain is released a second time at the end of tc_new_tfilter(). Always set
'tp_created' flag when trying to insert new proto.

Change-Id: I98543720330df217556652ad278d044a42a5af56
Signed-off-by: Vlad Buslov <vladbu@mellanox.com>